freakvpn-bot/requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
apscheduler==3.10.4
pytz==2023.3
python-dotenv==1.0.0
aiosqlite==0.19.0
loguru==0.7.2
